Fixes to bits.cpp, debug.cpp, more unit tests

+++ support/bits.cpp
@@ -18,7 +18,7 @@ int trailing_zeroes(uint32_t v)
 		31, 27, 13, 23, 21, 19, 16, 7, 26, 12, 18, 6, 11, 5, 10, 9
 	};
 	
-	return MultiplyDeBruijnBitPosition[((v & -v) * 0x077CB531UL) >> 27];
+	return MultiplyDeBruijnBitPosition[((uint32_t)((v & -v) * 0x077CB531UL)) >> 27];
 #endif
 }
 
@@ -39,7 +39,7 @@ int bottom_bit(uint32_t v)
 		31, 27, 13, 23, 21, 19, 16, 7, 26, 12, 18, 6, 11, 5, 10, 9
 	};
 	
-	return MultiplyDeBruijnBitPosition[((v & -v) * 0x077CB531UL) >> 27];
+	return MultiplyDeBruijnBitPosition[((uint32_t)((v & -v) * 0x077CB531UL)) >> 27];
 #endif
 }
